Reorganize models and update adapter imports
