+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/models/registry_farmer.py
@@ -1,18 +1,14 @@
-from datetime import datetime
+# from datetime import datetime
 
-from openg2p_fastapi_common.models import BaseORMModel
-from sqlalchemy import DateTime, Float, Integer, String
+# from openg2p_fastapi_common.models import BaseORMModel
+from openg2p_pbms_models.models import G2PRegistry
+from sqlalchemy import Float, Integer, String
 from sqlalchemy.orm import mapped_column
 
 
-class G2PFarmerRegistry(BaseORMModel):
+class G2PFarmerRegistry(G2PRegistry):
     __tablename__ = "g2p_farmer_registry"
 
-    id = mapped_column(Integer, primary_key=True, autoincrement=True)
-    unique_id = mapped_column(String, nullable=True)
-    registration_date = mapped_column(
-        DateTime, default=datetime.utcnow(), nullable=False
-    )
     name = mapped_column(String, nullable=False)
     land_area = mapped_column(Float, nullable=True)
     no_of_cattle_heads = mapped_column(Integer, nullable=True)
