+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/models/registry_student.py
@@ -1,18 +1,14 @@
-from datetime import datetime
+# from datetime import datetime
 
-from openg2p_fastapi_common.models import BaseORMModel
-from sqlalchemy import Date, DateTime, Integer, String
+# from openg2p_fastapi_common.models import BaseORMModel
+from openg2p_pbms_models.models import G2PRegistry
+from sqlalchemy import Date, String
 from sqlalchemy.orm import mapped_column
 
 
-class G2PStudentRegistry(BaseORMModel):
+class G2PStudentRegistry(G2PRegistry):
     __tablename__ = "g2p_student_registry"
 
-    id = mapped_column(Integer, primary_key=True, autoincrement=True)
-    unique_id = mapped_column(String, nullable=True)
-    registration_date = mapped_column(
-        DateTime, default=datetime.utcnow(), nullable=False
-    )
     name = mapped_column(String, nullable=False)
     institution_name = mapped_column(String, nullable=True)
     date_of_birth = mapped_column(Date, nullable=True)
