+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/models/registry_type.py
