+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/schema/__init__.py
@@ -1,5 +1,6 @@
+from .eligibility_summary import EligibilitySummaryResponse
 from .eligibility_summary_farmer import EligibilitySummaryFarmerResponse
 from .eligibility_summary_student import EligibilitySummaryStudentResponse
-from .registry import G2PRegistry
-from .registry_farmer import G2PFarmerRegistry
-from .registry_student import G2PStudentRegistry
+from .registry import G2PRegistryResponse
+from .registry_farmer import G2PFarmerRegistryResponse
+from .registry_student import G2PStudentRegistryResponse
