+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/schema/registry.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from pydantic import BaseModel
 
 
-class G2PRegistry(BaseModel):
+class G2PRegistryResponse(BaseModel):
     id: int
     unique_id: str
     registration_date: datetime
