+++ openg2p-eee-registry-adapters/src/openg2p_eee_registry_adapters/schema/registry_student.py
@@ -1,10 +1,10 @@
 from datetime import datetime
 from typing import Optional
 
-from .registry import G2PRegistry
+from .registry import G2PRegistryResponse
 
 
-class G2PStudentRegistry(G2PRegistry):
+class G2PStudentRegistryResponse(G2PRegistryResponse):
     name: str
     institution_name: Optional[str]
     date_of_birth: Optional[datetime]
