+++ openg2p-pbms-models/src/openg2p_pbms_models/models/__init__.py
@@ -1,2 +1,3 @@
 from .program_definiton import G2PProgramDefinition
 from .que_eee_request import EnumStatus, G2PQueEEERequest
+from .registry import G2PRegistry
